+++ Project3-OverlayNetwork/overlay.h
@@ -25,9 +25,9 @@
 void printUsage();
 int runRouter(char* ipMappings);
 int runEndHost(char* routerIP, char* hostIP, uint32_t ttl);
-int routePacket(char* packet, std::vector<std::string> &overlayIPs, std::vector<std::string> &vmIPs);
+int8_t routePacket(char* packet, std::vector<std::string> &overlayIPs, std::vector<std::string> &vmIPs);
 void readUDP(char* packet);
 void sendUDP(char* routerIP,char* sourceaddr, char* destaddr, uint32_t ttl);
-std::string retcodeString(int retcode);
-void routerlog(std::string srcOverlayIP, std::string dstOverlayIP, unsigned short ipdent, int retcode);
+std::string retcodeString(int8_t retcode);
+void routerlog(std::string srcOverlayIP, std::string dstOverlayIP, unsigned short ipdent, int8_t retcode);
 bool lookForFileAndProcess();
